+++ src/components/pages/PageThreadShow.vue
@@ -1,34 +1,25 @@
 <template>
-    <div class="col-large push-top">
-      <h1>{{ thread.title }}</h1>
-      <PostList
-        :posts="posts"
-      />
-      <form @submit.prevent="addPost">
-        <div class="form-group">
-          <textarea 
-            name="" 
-            id="" 
-            cols="30"
-            rows="10"
-            class="form-input"
-            v-model="newPostText"
-          ></textarea>
-        </div>
-        <div class="form-action">
-          <button class="btn-blue">Submit Post</button>
-        </div>
-      </form>
-    </div>
+  <div class="col-large push-top">
+    <h1>{{ thread.title }}</h1>
+    <PostList
+      :posts="posts"
+    />
+    <PostEditor
+     @save="addPost"
+     :threadId="id"
+    />
+  </div>
 </template>
 
 <script>
 import sourceData from '@/data'
 import PostList from '@/components/PostList'
+import PostEditor from '@/components/PostEditor'
 
 export default {
   components: {
-    PostList
+    PostList,
+    PostEditor
   },
 
   props: {
@@ -40,8 +31,7 @@ export default {
 
   data () {
     return {
-      thread: sourceData.threads[this.id],
-      newPostText: ''
+      thread: sourceData.threads[this.id]
     }
   },
 
@@ -54,19 +44,12 @@ export default {
   },
 
   methods: {
-    addPost () {
-      const postId = 'greatPost' + Math.random()
-      const post = {
-        text: this.newPostText,
-        publishedAt: Math.floor(Date.now() / 1000),
-        threadId: this.id,
-        userId: '7uVPJS9GHoftN58Z2MXCYDqmNAh2',
-        '.key': postId
-      }
+    addPost (evenData) {
+      const post = evenData.post
+      const postId = evenData.post['.key']
       this.$set(sourceData.posts, postId, post)
       this.$set(this.thread.posts, postId, postId)
       this.$set(sourceData.users[post.userId].posts, postId, postId)
-      this.newPostText = ''
     }
   }
 }
